add v2/tests to hold int tests; add units
